etl_script/requirements.txt
httpx==0.27.*
pandas==2.2.*
sqlalchemy==2.0.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*